requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
